Fixed indentation problems in the _create_sample_exogenous_chart function - there was extra indentation which was causing syntax errors.
Corrected the missing newline between function definitions - there was a syntax error where _process_real_exogenous_data was defined immediately after the closing parenthesis of the previous function without a proper newline.

Fixed the implementation of rx.cond() in create_exogenous_chart - when using reactive conditionals with functions that take arguments, we need to wrap them in lambda functions to ensure they are only evaluated when needed.

Removed duplicate code - there was a redundant return rx.box(...) statement at the end of the file that was unreachable.